Change Aws security groups to not delete default rules and add security groups on profiles.

+++ code/environments/production/modules/aws/lib/puppet/provider/ec2_securitygroup/v2.rb
@@ -210,9 +210,11 @@ Puppet::Type.type(:ec2_securitygroup).provide(:v2, :parent => PuppetX::Puppetlab
 
   def destroy
     Puppet.info("Deleting security group #{name} in region #{target_region}")
-    ec2.delete_security_group(
-      group_id: @property_hash[:id]
-    )
-    @property_hash[:ensure] = :absent
+    if name != "default"
+      ec2.delete_security_group(
+        group_id: @property_hash[:id]
+      )
+      @property_hash[:ensure] = :absent
+    end
   end
 end
